requirements.txt
pandas==2.3.0
python-dateutil==2.9.0.post0
seaborn
matplotlib
tqdm
sentence-transformers
langchain
faiss-cpu
python-dotenv
langchain-core
langchain-text-splitters